convert tests to jest
